Replaced Toast with Snackbar for password validation message
Issue: The Toast message for password validation was being cut off due to its length, making it difficult for users to read the entire message.

Fix: Replaced the Toast message with a Snackbar in the SignUp fragment. Snackbar provides more flexibility than Toast, including multiline text and action to interact with. This ensures the entire validation message is visible to the users."